Revert "Feature/add filter"
