+++ backend/package.json
@@ -7,12 +7,8 @@
   "private": true,
   "dependencies": {
     "apollo-server": "^3.2.0",
-    "firebase": "^9.6.7",
-    "firebase-admin": "^10.0.2",
     "graphql": "^15.5.1",
-    "mysql2": "^2.3.3",
     "nodemon": "^2.0.12",
-    "sequelize": "^6.16.2",
     "ts-node": "^10.2.1",
     "typescript": "^4.3.5"
   },
