--- backend/serviceAccountKey.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "feather-b5d2a",
-  "private_key_id": "dfa4ff0c24c35125b8c65af99e420c9b193ae116",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCzlmAgoqco0veT\ndvGVBM2r2XxszV5cjqJ0mPP/rz8DtGWm3QFpbSQrsSgVN0Xzc8KOg9DbQwR/Zamv\nic0P02pmuI0qa+ILtauvMYpuBfHUx6iQVimDyMxK9/To3DRZQL2AlY9eo1RmOr14\n/HP2yoFhhFCUx9D2jR+0GgbhjafyYgs0iGnhdKZir/FDzAB1/n2CrvOpYSeUmXdk\nMxmtDfiTo76THz2WjUDFyB/C1rx/Zt2Ni5u+l1tGXYL4UKwaunib8pHYLBpKlMN0\nF+Swut9EZ3WuZzwmbMHGrO86WT2yhGtB9ewVJoB1Yb4wccQjQ+b8/2rr45a5/ZT0\nqWbCKzhTAgMBAAECggEALNl+MkhMtM5TFBndZVjowGU1RLrZOQZKHhJgxVcPxq4T\nwONqydh+8wZwOzQtnQoo6YyT6V5ITkSmZvdkJfKnnPW0E4cT0dbSBJO7vZRwe7JN\nL5bFEWthUcuw90KfNDcN3IH9ZyYlGs7KqN9rWpPiUedIdI+eaSR/aWwikcvdyJM2\nnKmuQdFZJ31QrW5K3FraYltNa2JQvK2iiFb8ZInj5KA1yKy6BtL+4SAXyzEqu7A1\na5PAT646T+1Yv1tRrZVTuhxD607JCo3ZPRP5hcgEpxjKlv/w62sp+FA+r9bvEW+v\njNWyuoqZZCH+MUrWmksPoZcerM7zISIWPWe8qnok8QKBgQDY0IwJ3gOBbMoR0pQt\neNZHpAqbhORndRv/hLGWyLIbap24IHzvAYQLOuqKtCUbI45hSgXsqQaGwGnpTzWJ\nAm/ko3nqNPYyCQJqJVMTsOqKGgFTbwtOMZEMHtijCAKFxVfQlWDON1mP9N7XBAFA\n06dt58cnVn2Qjw4ptrqTLEsbKQKBgQDUC2ut5iLwTI8Ds92l3pU732XMbKD5QrFU\nzkrdclBwZhHT8a4C4rnBn5cyr2x7SlAPl4NO5kodxMmn0g2/ZjNAiI71cZXoXD3f\nssKF/N1i5ZYUPHkKAt8ZAE7wxbdh+NyxgJBuXjm9rBTwqdGxi12wtWPOFnnPrCDe\n4q6uQxrjGwKBgQCnufWLsJePqSRNU1VUNBW4iDu4gb7cFimZqwJjPlEiioEGrwNX\nNhn7QhGtSZ0m+IOt2cDpivzEg9erKJ6mph7/hkPLHjuTC4s1qDgzlHpsaZ+69r2N\nOSUXdklrHcgtyjOb5eNkxEbh6bH2Ll/10YbFrDdRxEQQ78oqFPUS44IauQKBgEUD\nWo6EwoZXRIy987vSj9umv+kiydmWnUSux6r/U7LhZZgHps0FodOk737tnW8bJqeM\ndo4tYZ4CuGdIdHUA/FlamLzFtTalHvY5phoZzKwtcQUxEWvyt03ZzdeLYj4e0yVq\nZJenTz+uTDpv6vHQHDQxNUSUCB1RNnmUaaoJ6SxVAoGBAJHfQZiquAp8n53DWKFE\nNcqvDkSGDOoviNVkvFWwftIpcihvpHvq9LpWl51dwCR/EU5cq4C2WEro7a/SUck3\nQHSmZFICA250lZoGSC2ZOpCbHLnDZV8WiXNQ1Owgjoa0jMzsSS68UVwBonynRwTI\npms74/H9F8tDlDlIyjhFkvyw\n-----END PRIVATE KEY-----\n",
-  "client_email": "firebase-adminsdk-enust@feather-b5d2a.iam.gserviceaccount.com",
-  "client_id": "109478194477987205470",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-enust%40feather-b5d2a.iam.gserviceaccount.com"
-}
