--- backend/src/config/firebase.ts
@@ -1,8 +1,0 @@
-export const firebaseConfig = {
-  apiKey: "AIzaSyCrVsifbaDH0rWCK-JiPgmcDTUT_Cvpx80",
-  authDomain: "feather-b5d2a.firebaseapp.com",
-  projectId: "feather-b5d2a",
-  storageBucket: "feather-b5d2a.appspot.com",
-  messagingSenderId: "21260938239",
-  appId: "1:21260938239:web:65e1803fa2fad5aabf55ca",
-};
