+++ backend/src/dateScalar.ts
@@ -6,6 +6,13 @@ const dateScalar = new GraphQLScalarType({
   parseValue(value) {
     return new Date(value); // Convert incoming integer to Date
   },
+  parseLiteral(ast) {
+    console.log(ast);
+    if (ast.kind === Kind.INT) {
+      return new Date(parseInt(ast.value, 10)); // Convert hard-coded AST string to integer and then to Date
+    }
+    return null; // Invalid hard-coded value (not an integer)
+  },
   serialize(value) {
     const date = new Date(value);
