+++ backend/src/mockData.ts
@@ -27,88 +27,4 @@ export const policies = [
     endDate: new Date(),
     createdAt: 123232,
   },
-  {
-    customer: {
-      firstName: "John",
-      lastName: "Leguizamo",
-      dateOfBirth: new Date(2020, 12, 12),
-    },
-    provider: "Allianz",
-    insuranceType: "LIABILITY",
-    status: "ACTIVE",
-    policyNumber: "12334",
-    startDate: new Date(),
-    endDate: new Date(),
-    createdAt: new Date(),
-  },
-  {
-    customer: {
-      firstName: "Chris",
-      lastName: "Pratt",
-      dateOfBirth: new Date(2020, 12, 12),
-    },
-    provider: "Allianz",
-    insuranceType: "LIABILITY",
-    status: "ACTIVE",
-    policyNumber: "12334",
-    startDate: new Date(),
-    endDate: new Date(),
-    createdAt: new Date(),
-  },
-  {
-    customer: {
-      firstName: "Ed",
-      lastName: "Sheeran",
-      dateOfBirth: new Date(2019, 1, 18),
-    },
-    provider: "La Caja",
-    insuranceType: "LIABILITY",
-    status: "ACTIVE",
-    policyNumber: "12334",
-    startDate: new Date(),
-    endDate: new Date(),
-    createdAt: new Date(),
-  },
-  {
-    customer: {
-      firstName: "Ozzy",
-      lastName: "Osbourne",
-      dateOfBirth: new Date(2020, 12, 12),
-    },
-    provider: "Apollo",
-    insuranceType: "LIABILITY",
-    status: "ACTIVE",
-    policyNumber: "12334",
-    startDate: new Date(),
-    endDate: new Date(),
-    createdAt: new Date(),
-  },
-  {
-    customer: {
-      firstName: "Paul",
-      lastName: "Newman",
-      dateOfBirth: new Date(2020, 12, 12),
-    },
-    provider: "Barmer",
-    insuranceType: "LIABILITY",
-    status: "ACTIVE",
-    policyNumber: "12334",
-    startDate: new Date(),
-    endDate: new Date(),
-    createdAt: new Date(),
-  },
-  {
-    customer: {
-      firstName: "Will",
-      lastName: "Smith",
-      dateOfBirth: new Date(2020, 12, 12),
-    },
-    provider: "TK",
-    insuranceType: "LIABILITY",
-    status: "ACTIVE",
-    policyNumber: "12334",
-    startDate: new Date(),
-    endDate: new Date(),
-    createdAt: new Date(),
-  },
 ];
