--- backend/src/models/customer.model.ts
@@ -1,1 +1,0 @@
-import { Model, DataTypes } from "sequelize";
