+++ frontend/package.json
@@ -5,7 +5,6 @@
   "dependencies": {
     "@apollo/client": "^3.5.9",
     "@craco/craco": "^6.2.0",
-    "@popsure/dirty-swan": "^0.26.7",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
@@ -14,11 +13,9 @@
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "graphql": "^16.3.0",
-    "mysql2": "^2.3.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
-    "sequelize": "^6.16.2",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
