--- frontend/src/interfaces/filteredPolicies.ts
@@ -1,22 +1,0 @@
-// Generated by https://quicktype.io
-
-export interface FilteredPolicies {
-  filterPolicies: FilterPolicy[];
-}
-
-export interface FilterPolicy {
-  customer: Customer;
-  provider: string;
-  insuranceType: string;
-  policyNumber: string;
-  status: string;
-  startDate: string;
-  endDate: string;
-  createdAt: string;
-}
-
-export interface Customer {
-  firstName: string;
-  lastName: string;
-  dateOfBirth: string;
-}
